item_uploader/requirements.txt
gspread
python-dotenv
streamlit
pandas
gspread-formatting
xlsxwriter